authrizion is working now

+++ src/tasks/tasks.service.ts
@@ -7,6 +7,8 @@ import { TaskRepository } from './task.repository';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Task } from './task.entity';
 import { TaskStatus } from './task-status.enum';
+import { User } from '../auth/user.entity';
+import { UpdateResult } from 'typeorm';
  @Injectable()
 export class TasksService {
      constructor(
@@ -14,53 +16,66 @@ export class TasksService {
         private taskRepository: TaskRepository
     ){}
 
-       async getTasks(filterDto:GetTaskFilterDto):Promise<Task[]>{
-            const {status ,search} = filterDto
-            const query = this.taskRepository.createQueryBuilder('task')//refer to task entity
-            if(status){
-                query.andWhere('task.status = :status',{status})
-            }
-            if(search){
-                query.andWhere('(task.title LIKE :search OR task.description LIKE :search)',{search: `%search%`})// if we want to search for part of the title we have to use %%
-            }
-            const tasks = await query.getMany()
-            return tasks 
+       async getTasks(filterDto:GetTaskFilterDto
+        , user:User 
+        ):Promise<Task[]>{
+            // const {status ,search} = filterDto
+            // const query = this.taskRepository.createQueryBuilder('task')//refer to task entity
+            // if(status){
+            //     query.andWhere('task.status = :status',{status})
+            // }
+            // if(search){
+            //     query.andWhere('(task.title LIKE :search OR task.description LIKE :search)',{search: `%search%`})// if we want to search for part of the title we have to use %%
+            // }
+            // const tasks = await query.getMany()
+            return  this.taskRepository.getTasks(filterDto,user) 
         }    
 
-        async getTaskById(id:number):Promise<Task>{
-            const found = await this.taskRepository.findOne(id)
+        async getTaskById(id:number,
+            user:User
+            ):Promise<Task>{
+           
+                     const found = await this.taskRepository.findOne({where:{id ,userId:user.id}} )
  
-                if(!found){
-                 throw new NotFoundException(`Task with ID ${id} is Not found`);
-                              
-                }
-                    return   found
+                    if(!found){
+                     throw new NotFoundException(`Task with ID ${id} is Not found`);
+                                  
+                    }
+                        return   found
            }
                   
              //    return await this.taskRepository.createTask(createTaskDto);
-            async createTask(createTaskDto:CreateTaskDto):Promise<Task>{
-                const {title,description} = createTaskDto ;
-                const task = new Task() ;
-                task.title = title ;
-                task.description = description;
-                task.status = TaskStatus.Done ;
-               await this.taskRepository.save(task) ;
-               return task;
+            async createTask(createTaskDto:CreateTaskDto,
+                 user:User 
+                ):Promise<Task>{
+            //     const {title,description} = createTaskDto ;
+            //     const task = new Task() ;
+            //     task.title = title ;
+            //     task.description = description;
+            //     task.status = TaskStatus.Done ;
+            //    await this.taskRepository.save(task) ;
+            //    return task;
+               return this.taskRepository.createTask(createTaskDto,user)
            }
-        async deleteTask(id:number):Promise<void>{
-            const result = await this.taskRepository.delete(id)
+        async deleteTask(id:number,
+            user:User
+            ):Promise<void>{
+            const result = await this.taskRepository.delete({id,userId:user.id})
             if(result.affected === 0){
                 throw new NotFoundException(`Task with ID ${id} is Not found`);
 
             }
             
         }
-        async updateTask(id:number , status: TaskStatus):Promise<Task>{
-            const found = await this.getTaskById(id)
+        async updateTask(id:number , status: TaskStatus,
+            user:User
+           ):Promise<Task>{
+            const found = await this.getTaskById(id,user)
             found.status = status 
             await found.save()
             return found
         }
+    }
 //     getAllTasks(): Task[] {
 //     return this.tasks ; 
 //     }
@@ -110,4 +125,3 @@ export class TasksService {
 //         this.tasks = this.tasks.filter(task => task.id !== found.id)
           
 //     }
-}
